Grass working with water mask, however it is too unoptimised currently. Will work on building mask tomorrow
