--- Assets/ProceduralPipeline/Nodes/Buildings/CreateChunksNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb839b70d3d731d49a6745ce5cb7ca32
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
