--- Assets/ProceduralPipeline/Nodes/Terrain/CreateGrassNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c4e347f8c2901564988bde3a0b205528
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
