+++ Assets/ProceduralPipeline/Nodes/Terrain/Grass/RemoveGrassOnWaterNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a24e9a456c08dc24bbadba117cc215d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
